+++ sim/rebforces/profiles.c
@@ -3,7 +3,7 @@
 
 #include "profiles.h"
 
-uint_fast8_t STD_PROF_N = 0;
+unsigned int STD_PROF_N = 0;
 
 double STD_PROF_X[STD_PROF_NMAX] = { NAN };
 double VELOCITY_PROF[2][STD_PROF_NMAX] = { { NAN } };
